Re-implementing the iterator_move like logic
The problem was my naive re-implementation of the iterator /
iterator_move logic! I naively assumed the iterator was simply
iterating NEXT each time but I failed to realize that the random seek
was also seeking to a specific key.

The fix is to re-implement the iterator_move function but write it to
operate on a list of key-values *like it was iterating over the
keyspace in leveldb*. This allows the iterator, when building the
hashtree, to jump around.

I maintained the PREFETCH and PREFETCH_STOP semantics to keep it as
close to the original as possible, even though it's just the NEXT
operation.